+++ resources/js/components/form/InputCheckbox.vue
@@ -2,6 +2,7 @@
   <div class="form-group row">
     <label :for="name | kebabCase" class="col-md-3 col-form-label text-md-right py-0">{{ label }}</label>
     <div class="col-md-6">
+      <input-hidden :id="name + '-unchecked-value' | kebabCase" :name="name" value="0"></input-hidden>
       <b-form-checkbox
         type="checkbox"
         :id="name | kebabCase"
@@ -10,7 +11,9 @@
         value="1"
         unchecked-value="0"
         :state="errorMessage ? false : null"
-        v-model="currentValue"></b-form-checkbox>
+        v-model="currentValue"
+        :switch="isSwitch"
+        :size="size"></b-form-checkbox>
 
       <span v-if="errorMessage" class="invalid-feedback" role="alert">
         <strong>{{ errorMessage }}</strong>
@@ -26,6 +29,13 @@ export default {
   mixins: [ Input ],
   props: {
     label: { type: String, required: true },
+    switch: { type: Boolean, default: false },
+    size: { type: String },
+  },
+  data() {
+    return {
+      isSwitch: this.switch
+    }
   }
 }
 </script>