+++ resources/js/components/quali/tiptap-extensions/observation/ModalAddObservation.vue
@@ -15,6 +15,9 @@
       show-content
       show-block
       show-user
+      :show-requirements="showRequirements"
+      :show-categories="showCategories"
+      :show-impression="showImpression"
       pointer-cursor
       @clickObservation="observation => select(observation.pivot.id)"></observation-list>
 
@@ -30,7 +33,10 @@
     props: {
       value: { type: Function, required: false },
       observations: { type: Array, required: true },
-      returnFocus: { type: Object, required: false }
+      returnFocus: { type: Object, required: false },
+      showRequirements: { type: Boolean, default: false },
+      showCategories: { type: Boolean, default: false },
+      showImpression: { type: Boolean, default: false },
     },
     inject: [ 'courseId', 'requirements', 'categories' ],
     computed: {