+++ resources/views/qualiContent/edit.blade.php
@@ -13,7 +13,10 @@
             :quali-contents="{{ json_encode($quali->contents) }}"
             :observations="{{ json_encode($observations) }}"
             :requirements="{{ json_encode($course->requirements) }}"
-            :categories="{{ json_encode($course->categories) }}">
+            :categories="{{ json_encode($course->categories) }}"
+            :show-requirements="{{ $course->uses_requirements ? 'true' : 'false' }}"
+            :show-categories="{{ $course->uses_categories ? 'true' : 'false' }}"
+            :show-impression="{{ $course->uses_impressions ? 'true' : 'false' }}">
 
             <b-button variant="link" class="px-0" href="{{ route('participants.detail', ['course' => $course->id, 'participant' => $participant->id]) }}">
                 <i class="fas fa-arrow-left"></i> {{__('t.views.quali_content.back_to_participant', ['name' => $participant->scout_name])}}