+++ tests/Feature/App/SelectLocaleWhenLoggedInTest.php
@@ -97,7 +97,7 @@ class SelectLocaleWhenLoggedInTest extends TestCase {
         $response = $this->get('/user', [], $headers);
 
         // then
-        $response->assertStatus(200);
+        $response->assertOk();
         $this->assertSeeAllInOrder('#navbar-locale-select', [$expected], function($domElement) { return $domElement->getAttribute('text'); });
         $this->assertSeeAllInOrder('#navbar-locale-select b-dropdown-item', $unselected);
         $response->assertSessionHas('locale', $expected);