+++ tests/Feature/Observation/UpdateObservationTest.php
@@ -52,7 +52,7 @@ class UpdateObservationTest extends TestCaseWithBasicData {
         $response->assertStatus(302);
         $response->assertRedirect(route('admin.course', ['course' => $this->courseId]));
     }
-
+  
     public function test_shouldDisplayAllBlocksInChronologicalOrder() {
         // given
         $this->createBlock('old block', 1.1, Carbon::now()->subDays(2)->format('d.m.Y'));
@@ -65,6 +65,55 @@ class UpdateObservationTest extends TestCaseWithBasicData {
 
         // then
         $response->assertSeeInOrder(['old block', 'yesterday', 'today', 'tomorrow']);
+    }
+
+    public function test_shouldDisplayOptionalFields_whenActivated() {
+        // given
+
+        // when
+        $response = $this->get('/course/' . $this->courseId . '/observation/' . $this->observationId);
+
+        // then
+        $response->assertOk();
+        $response->assertSee('label="Anforderungen"', false);
+        $response->assertSee('label="Eindruck"', false);
+        $response->assertSee('label="Kategorien"', false);
+    }
+
+    public function test_shouldNotDisplayRequirementsSelect_whenNoRequirementsInCourse() {
+        // given
+        Course::find($this->courseId)->requirements()->delete();
+
+        // when
+        $response = $this->get('/course/' . $this->courseId . '/observation/' . $this->observationId);
+
+        // then
+        $response->assertOk();
+        $response->assertDontSee('label="Anforderungen"', false);
+    }
+
+    public function test_shouldNotDisplayImpressionInput_whenDeactivatedInCourse() {
+        // given
+        Course::find($this->courseId)->update(['uses_impressions' => false]);
+
+        // when
+        $response = $this->get('/course/' . $this->courseId . '/observation/' . $this->observationId);
+
+        // then
+        $response->assertOk();
+        $response->assertDontSee('label="Eindruck"', false);
+    }
+
+    public function test_shouldNotDisplayCategorySelect_whenNoCategoriesInCourse() {
+        // given
+        Course::find($this->courseId)->categories()->delete();
+
+        // when
+        $response = $this->get('/course/' . $this->courseId . '/observation/' . $this->observationId);
+
+        // then
+        $response->assertOk();
+        $response->assertDontSee('label="Kategorien"', false);
     }
 
     public function test_shouldUpdateObservation() {
@@ -222,19 +271,22 @@ class UpdateObservationTest extends TestCaseWithBasicData {
         $this->assertEquals('Beobachtung darf maximal 1023 Zeichen haben.', $exception->validator->errors()->first('content'));
     }
 
-    public function test_shouldValidateNewObservationData_noImpression() {
+    public function test_shouldValidateNewObservationData_noImpression_shouldLeavePreviousImpression() {
         // given
         $payload = $this->payload;
         unset($payload['impression']);
+        Observation::find($this->observationId)->update(['impression' => 2]);
 
         // when
         $response = $this->post('/course/' . $this->courseId . '/observation/' . $this->observationId, $payload);
 
         // then
-        $this->assertInstanceOf(ValidationException::class, $response->exception);
-        /** @var ValidationException $exception */
-        $exception = $response->exception;
-        $this->assertEquals('Eindruck muss ausgefüllt sein.', $exception->validator->errors()->first('impression'));
+        $response->assertStatus(302);
+        /** @var TestResponse $response */
+        $response = $response->followRedirects();
+        $response->assertOk();
+        $this->assertNull($response->exception);
+        $this->assertEquals(2, Observation::find($this->observationId)->toArray()['impression']);
     }
 
     public function test_shouldValidateNewObservationData_invalidImpression() {